I've been working through the UI code and figuring out the whole structure.  I get most of it now and can probably start making some meaningful changes soon.  For the time being, I've done little things like clean up the code in places to remove warnings, to take advantage of more sane code structure, to add little nice-to-have features like Ctrl-Q now quits the program and Ctrl-O can open a file, etc.
This is my first time doing any development in NetBeans.  I checked it out a couple years ago and it really wasn't as good as Eclipse, but, thus far, it's pretty nice!  I love the visual GUI editor; makes it a lot easier to get work done.  For example, figuring out how to make the window resizable was a cinch.

git-svn-id: https://zoomwhee.mgh.harvard.edu/svn_devel/OpenMIMS/trunk@85 ea379df5-90a0-4b1b-ac45-2d7255ffb434