Renamed groupsMap to roisMap for clarity.
The renameGroupActionPerformed was not working correctly.   This was 
fixed.

Changing rois either one at a time or several at once from one group to 
another was severely broken.  That was also fixed.

In the getRoiGroup method, added a check for the retrieved group being 
null, and if so, set the group to the default group.  This was necessary 
because the default group is never added to the roisMap, a decision made 
by a prior programmer.

In addToGroupWithoutRenaming method, fixed an improper put to the 
roisMap.

In the roiPixelvalues method, added a check for group being null, and if 
so, added “default” to lgroups.

Extensive changes to the openZip method.

In the measure method, added a warning dialog if the user tried to do a 
measure if no roi is selected.

Fixed broken put to the roisMap in the duplicate method.

In several of the methods invoked from the “more” menu of the 
RoiManager2, such as intersection, complement, and exclusiveToRoi, 
added information dialogs to show results rather than just putting the 
results on the console with a System.out.println.

In the add method, added code to allow it to add multiple rois,  
including rois present in the default group.