Added a label near the bottom of the ROI manager window that counts down to the time in seconds until the next autosave.  If autosave is not active, this label in not visible.  When the countdown is in progress, the user can right-click on the label to bring up a menu that can be used to cancel the upcoming autosave event.    To implement this, a class that extends TimerTask was added (RemindTask).  There were also several methods added that work together with RemindTask to implement a 1 second timer (Reminder, cancelTimer, and startTimer).
For easier debugging, all places that formerly set or cleared 
needsToBeSaved directly now do it by calling a new method 
(setNeedsToBeSaved).

Minor bugs in the change group section of the huge roisHandler method 
were fixed.

The window was widened slightly so that the “sec” at the end of the 
timer would not be cut off on Linux.

A new listener was added to call jLabelNextAutosaveMouseClicked when 
the user right clicks on the label that shows the time of the last 
autosave.  This allows the user to do an immediate autosave, instead of
 waiting for the autosave interval timer to count down to zero.
Added lines to the end of saveMultiple to reset the timer.

The delete method has a new parameter (noDeleteAll) that prevents the 
user from accidentally deleting all of the ROIs when actually attempting 
to delete a single ROI using a new feature:  holding down the shift key 
and clicking on a selected ROI in an image window deletes that ROI.