In the constructor, there is a drag and drop section for opening files via drag and drop.  Unlike the section that opens files with the File -> Open menu, there were several actions present in the File -> open code that were missing the the drag and drop block.  There were calls to various methods in the ROI manager (selectAll, delete, close, and setNeedsToBeSave).   Adding these calls to the drag and drop section results in more consistent behavior between these two techniques.
Also in the drag and drop section, a call to a new method 
startAutosaveROIthread was added so that a new countdown to the next 
autosave could be created.

In openMIMSImageMenuItemActionPerformed, the autosave countdown thread 
is started.

In the crazy long restoreState method, in the code block that starts 
with if(!usedForTables), I disabled DJ’s call to addShownRatiosToList 
and enabled the call to hsiControl.updateImage(true).   This prevents 
unwanted rows from appearing the list on the left size of the Process 
tab window when the user opens a second file after opening an initial 
file.  It also prevent corruption of some of the HSI and/or ratio images 
in the same situation.  I am not certain what addShownRatiosToList was 
put in there to do, so I may have to put it back in later.