Added JavaDoc for method saveRoiToZipNow.
Added code to the run method to allow the thread to be interrupted.

In saveROItoZipNow, added logic to detect and deal with an issue in 
which sometimes the currentTime string is missing the seconds.
If a save is going to occur, the method now hides the autosave countdown 
in the ROI manager window until it finishes saving the ROI zip file.